Move object_store into s3_store.
