+++ pixeltable/catalog/table_version.py
@@ -20,7 +20,7 @@ from pixeltable.iterators import ComponentIterator
 from pixeltable.metadata import schema
 from pixeltable.utils.exception_handler import run_cleanup_on_exception
 from pixeltable.utils.filecache import FileCache
-from pixeltable.utils.media_store import MediaDestination
+from pixeltable.utils.media_destination import MediaDestination
 
 from .tbl_ops import TableOp
